# output/go-echo-server/handlers/api_inst_group.go
package handlers
import (
    "github.com/GIT_USER_ID/GIT_REPO_ID/models"
    "github.com/labstack/echo/v4"
    "net/http"
)

// CreateInstGroup - 施設グループを作成する
func (c *Container) CreateInstGroup(ctx echo.Context) error {
    return ctx.JSON(http.StatusOK, models.HelloWorld {
        Message: "Hello World",
    })
}


// GetInstGroup - 施設グループ情報を返す
func (c *Container) GetInstGroup(ctx echo.Context) error {
    return ctx.JSON(http.StatusOK, models.HelloWorld {
        Message: "Hello World",
    })
}


// GetInstGroups - 施設グループ一覧を返す
func (c *Container) GetInstGroups(ctx echo.Context) error {
    return ctx.JSON(http.StatusOK, models.HelloWorld {
        Message: "Hello World",
    })
}